Useless cast.

+++ vvp/compile.cc
@@ -17,7 +17,7 @@
  *    Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA
  */
 #ifdef HAVE_CVS_IDENT
-#ident "$Id: compile.cc,v 1.162 2003/05/24 02:48:37 steve Exp $"
+#ident "$Id: compile.cc,v 1.163 2003/05/25 03:04:55 steve Exp $"
 #endif
 
 # include  "arith.h"
@@ -1257,7 +1257,7 @@ void compile_code(char*label, char*mnem, comp_operands_t opa)
 		  // Make sure we dont overflow the short index
 		  if (opa->argv[idx].numb > 65535) {
 		      fprintf(stderr, "%lu overflows index for instruction %s\n",
-			  opa->argv[idx].numb, mnem);
+			      opa->argv[idx].numb, mnem);
 		      assert(0);
 		  }
 
@@ -1272,8 +1272,8 @@ void compile_code(char*label, char*mnem, comp_operands_t opa)
 
 		  // Make sure we dont overflow the short index
 		  if (opa->argv[idx].numb > 65535) {
-		      fprintf(stderr, "%d overflows index for instruction %s\n",
-			  (int)opa->argv[idx].numb, mnem);
+		      fprintf(stderr, "%lu overflows index for instruction %s\n",
+			      opa->argv[idx].numb, mnem);
 		      assert(0);
 		  }
 
@@ -1555,6 +1555,9 @@ void compile_param_string(char*label, char*name, char*str, char*value)
 
 /*
  * $Log: compile.cc,v $
+ * Revision 1.163  2003/05/25 03:04:55  steve
+ *  Useless cast.
+ *
  * Revision 1.162  2003/05/24 02:48:37  steve
  *  More thorough overflow error message.
  *
